+++ crates/transaction-pool/src/identifier.rs
@@ -1,9 +1,9 @@
-use crate::U256;
+
 use fnv::FnvHashMap;
 use reth_primitives::Address;
 use std::{
     collections::HashMap,
-    ops::{Bound, RangeBounds, RangeFrom},
+    ops::{Bound},
 };
 
 /// An internal mapping of addresses.
