+++ crates/transaction-pool/src/pool/mod.rs
@@ -67,43 +67,40 @@
 //!       transactions are _currently_ waiting for state changes that eventually move them into
 //!       category (2.) and become pending.
 use crate::{
-    error::{PoolError, PoolResult},
+    error::{PoolResult},
     pool::{
         listener::PoolEventListener,
-        pending::PendingTransactions,
-        queued::{QueuedPoolTransaction, QueuedTransactions},
     },
     traits::PoolTransaction,
     validate::ValidPoolTransaction,
     BlockID, PoolClient, PoolConfig, TransactionOrdering, TransactionValidator, U256,
 };
-use fnv::FnvHashMap;
+
 use futures::channel::mpsc::{channel, Receiver, Sender};
 use parking_lot::{Mutex, RwLock};
 use reth_primitives::{TxHash, U64};
 use std::{
-    collections::{HashMap, VecDeque},
+    collections::{HashMap},
     fmt,
     sync::Arc,
 };
-use tracing::{debug, trace, warn};
+use tracing::{warn};
 
+mod basefee;
 mod events;
 mod listener;
 mod pending;
-mod pending2;
 mod queued;
-mod queued2;
 pub(crate) mod state;
 mod transaction;
-pub mod txpool2;
+pub mod txpool;
 
 use crate::{
-    identifier::{SenderId, TransactionId},
     validate::TransactionValidationOutcome,
 };
 pub use events::TransactionEvent;
-pub use pending::TransactionsIterator;
+use crate::pool::pending::BestTransactions;
+use crate::pool::txpool::TxPool;
 
 /// Shareable Transaction pool.
 pub struct BasicPool<P: PoolClient, T: TransactionOrdering> {
@@ -209,7 +206,7 @@ pub struct PoolInner<P: PoolClient, T: TransactionOrdering> {
     /// Chain/Storage access.
     client: Arc<P>,
     /// The internal pool that manages
-    pool: RwLock<GraphPool<T>>,
+    pool: RwLock<TxPool<T>>,
     /// Pool settings.
     config: PoolConfig,
     /// Manages listeners for transaction state change events.
@@ -231,7 +228,7 @@ where
             client,
             config,
             event_listener: Default::default(),
-            pool: RwLock::new(GraphPool::new(ordering)),
+            pool: RwLock::new(TxPool::new(ordering)),
             ready_transaction_listener: Default::default(),
         }
     }
@@ -266,7 +263,7 @@ where
         tx: TransactionValidationOutcome<T::Transaction>,
     ) -> PoolResult<TxHash> {
         match tx {
-            TransactionValidationOutcome::Valid { balance, state_nonce, transaction } => {
+            TransactionValidationOutcome::Valid { balance: _, state_nonce: _, transaction: _ } => {
                 // TODO create `ValidPoolTransaction`
 
                 // let added = self.pool.write().add_transaction(tx)?;
@@ -334,221 +331,8 @@ where
     }
 
     /// Returns an iterator that yields transactions that are ready to be included in the block.
-    pub(crate) fn ready_transactions(&self) -> TransactionsIterator<T> {
-        self.pool.read().ready_transactions()
-    }
-}
-
-/// A pool that only manages transactions.
-///
-/// This pool maintains a dependency graph of transactions and provides the currently ready
-/// transactions.
-pub struct GraphPool<T: TransactionOrdering> {
-    /// How to order transactions.
-    ordering: Arc<T>,
-    /// Contains the currently known info
-    sender_info: FnvHashMap<SenderId, SenderInfo>,
-    /// Sub-Pool of transactions that are ready and waiting to be executed
-    pending: PendingTransactions<T>,
-    /// Sub-Pool of transactions that are waiting for state changes that eventually turn them
-    /// valid, so they can be moved in the `pending` pool.
-    queued: QueuedTransactions<T>,
-}
-
-// === impl PoolInner ===
-
-impl<T: TransactionOrdering> GraphPool<T> {
-    /// Create a new graph pool instance.
-    pub fn new(ordering: Arc<T>) -> Self {
-        let pending = PendingTransactions::new(Arc::clone(&ordering));
-        Self { ordering, sender_info: Default::default(), pending, queued: Default::default() }
-    }
-
-    /// Updates the pool based on the changed base fee.
-    ///
-    /// This enforces the dynamic fee requirement.
-    /// If the `new_base_fee` is _higher_ than previous base fee, all EIP-1559 transactions in the
-    /// ready queue that now violate the dynamic fee requirement need to parked.
-    /// If the `new_base_fee` is _lower_ than the previous base fee, all parked transactions that
-    /// now satisfy the dynamic fee requirement need to moved to the ready queue.
-    pub(crate) fn update_base_fee(&mut self, new_base_fee: U256) {
-        let _old_base_fee = self.pending.set_next_base_fee(new_base_fee);
-        // TODO update according to the changed base_fee
-        todo!()
-    }
-
-    /// Returns if the transaction for the given hash is already included in this pool
-    pub(crate) fn contains(&self, tx_hash: &TxHash) -> bool {
-        self.queued.contains(tx_hash) || self.pending.contains(tx_hash)
-    }
-
-    /// Returns an iterator that yields transactions that are ready to be included in the block.
-    pub(crate) fn ready_transactions(&self) -> TransactionsIterator<T> {
-        self.pending.get_transactions()
-    }
-
-    /// Adds the transaction into the pool
-    ///
-    /// This pool consists of two sub-pools: `Queued` and `Pending`.
-    ///
-    /// The `Queued` pool contains transaction with gaps in its dependency tree: It requires
-    /// additional transaction that are note yet present in the pool.
-    ///
-    /// The `Pending` pool contains all transactions that have all their dependencies satisfied (no
-    /// nonce gaps). It consists of two parts: `Parked` and `Ready`.
-    ///
-    /// The `Ready` queue contains transactions that are ready to be included in the pending block.
-    /// With EIP-1559, transactions can become executable or not without any changes to the
-    /// sender's balance or nonce and instead their feeCap determines whether the transaction is
-    /// _currently_ (on the current state) ready or needs to be parked until the feeCap satisfies
-    /// the block's baseFee.
-    fn add_transaction(
-        &mut self,
-        tx: ValidPoolTransaction<T::Transaction>,
-    ) -> PoolResult<AddedTransaction<T::Transaction>> {
-        if self.contains(tx.hash()) {
-            warn!(target: "txpool", "[{:?}] Already added", tx.hash());
-            return Err(PoolError::AlreadyAdded(Box::new(*tx.hash())))
-        }
-
-        let tx = QueuedPoolTransaction::new(tx, self.pending.provided_dependencies());
-        trace!(target: "txpool", "[{:?}] {:?}", tx.transaction.hash(), tx);
-
-        // If all ids are not satisfied import to queued
-        if !tx.is_satisfied() {
-            let hash = *tx.transaction.hash();
-            self.queued.add_transaction(tx)?;
-            return Ok(AddedTransaction::Queued { hash })
-        }
-        self.add_pending_transaction(tx)
-    }
-
-    /// Adds the transaction to the pending pool.
-    ///
-    /// This will also move all transaction that get unlocked by the dependency id this transaction
-    /// provides from the queued pool into the pending pool.
-    ///
-    /// CAUTION: this expects that transaction's dependencies are fully satisfied
-    fn add_pending_transaction(
-        &mut self,
-        tx: QueuedPoolTransaction<T>,
-    ) -> PoolResult<AddedTransaction<T::Transaction>> {
-        let hash = *tx.transaction.hash();
-        trace!(target: "txpool", "adding pending transaction [{:?}]", hash);
-        let mut pending = AddedPendingTransaction::new(hash);
-
-        // tracks all transaction that can be moved to the pending pool, starting the given
-        // transaction
-        let mut pending_transactions = VecDeque::from([tx]);
-        // tracks whether we're processing the given `tx`
-        let mut is_new_tx = true;
-
-        // take first transaction from the list
-        while let Some(current_tx) = pending_transactions.pop_front() {
-            // also add the transaction that the current transaction unlocks
-            pending_transactions
-                .extend(self.queued.satisfy_and_unlock(&current_tx.transaction.transaction_id));
-
-            let current_hash = *current_tx.transaction.hash();
-
-            // try to add the transaction to the ready pool
-            match self.pending.add_transaction(current_tx) {
-                Ok(replaced_transactions) => {
-                    if !is_new_tx {
-                        pending.promoted.push(current_hash);
-                    }
-                    // tx removed from ready pool
-                    pending.removed.extend(replaced_transactions);
-                }
-                Err(err) => {
-                    // failed to add transaction
-                    if is_new_tx {
-                        debug!(target: "txpool", "[{:?}] Failed to add tx: {:?}", current_hash,
-        err);
-                        return Err(err)
-                    } else {
-                        pending.discarded.push(current_hash);
-                    }
-                }
-            }
-            is_new_tx = false;
-        }
-
-        // check for a cycle where importing a transaction resulted in pending transactions to be
-        // added while removing current transaction. in which case we move this transaction back to
-        // the pending queue
-        if pending.removed.iter().any(|tx| *tx.hash() == hash) {
-            self.pending.clear_transactions(&pending.promoted);
-            return Err(PoolError::CyclicTransaction)
-        }
-
-        Ok(AddedTransaction::Pending(pending))
-    }
-
-    /// Prunes the transactions that provide the given dependencies.
-    ///
-    /// This will effectively remove those transactions that satisfy the dependencies.
-    /// And queued transactions might get promoted if the pruned dependencies unlock them.
-    pub fn prune_transactions(
-        &mut self,
-        dependencies: impl IntoIterator<Item = TransactionId>,
-    ) -> PruneResult<T::Transaction> {
-        let mut imports = vec![];
-        let mut pruned = vec![];
-
-        for dependency in dependencies {
-            // mark as satisfied and store the transactions that got unlocked
-            imports.extend(self.queued.satisfy_and_unlock(&dependency));
-            // prune transactions
-            pruned.extend(self.pending.remove_mined(dependency));
-        }
-
-        let mut promoted = vec![];
-        let mut failed = vec![];
-        for tx in imports {
-            let hash = *tx.transaction.hash();
-            match self.add_pending_transaction(tx) {
-                Ok(res) => promoted.push(res),
-                Err(e) => {
-                    warn!(target: "txpool", "Failed to promote tx [{:?}] : {:?}", hash, e);
-                    failed.push(hash)
-                }
-            }
-        }
-
-        PruneResult { pruned, failed, promoted }
-    }
-
-    /// Remove the given transactions from the pool.
-    pub fn remove_invalid(
-        &mut self,
-        tx_hashes: Vec<TxHash>,
-    ) -> Vec<Arc<ValidPoolTransaction<T::Transaction>>> {
-        // early exit in case there is no invalid transactions.
-        if tx_hashes.is_empty() {
-            return vec![]
-        }
-        trace!(target: "txpool", "Removing invalid transactions: {:?}", tx_hashes);
-
-        let mut removed = self.pending.remove_with_dependencies(tx_hashes.clone(), None);
-        removed.extend(self.queued.remove(tx_hashes));
-
-        trace!(target: "txpool", "Removed invalid transactions: {:?}", removed);
-
-        removed
-    }
-
-    /// Returns the current size of the entire pool
-    pub fn size_of(&self) -> usize {
-        unimplemented!()
-    }
-
-    /// Ensures that the transactions in the sub-pools are within the given bounds.
-    ///
-    /// If the current size exceeds the given bounds, the worst transactions are evicted from the
-    /// pool and returned.
-    pub fn enforce_size_limits(&mut self) {
-        unimplemented!()
+    pub(crate) fn ready_transactions(&self) -> BestTransactions<T> {
+        self.pool.read().best_transactions()
     }
 }
