--- crates/transaction-pool/src/pool/pending2.rs
@@ -1,256 +1,0 @@
-use crate::{
-    identifier::TransactionId,
-    pool::{
-        pending::PoolTransactionRef, queued::QueuedPoolTransaction,
-        txpool2::PoolInternalTransaction, TransactionsIterator,
-    },
-    PoolResult, TransactionOrdering, ValidPoolTransaction,
-};
-use fnv::FnvHashMap;
-use reth_primitives::rpc::TxHash;
-use std::{
-    cmp::Ordering,
-    collections::{BTreeMap, BTreeSet, HashMap, HashSet},
-    sync::Arc,
-};
-use tracing::debug;
-
-/// A pool of validated and gapless transactions that are ready on the current state and are waiting
-/// to be included in a block.
-pub(crate) struct PendingPool<T: TransactionOrdering> {
-    /// How to order transactions.
-    ordering: Arc<T>,
-    /// Keeps track of transactions inserted in the pool.
-    ///
-    /// This way we can determine when transactions where submitted to the pool.
-    id: u64,
-    /// All Transactions that are currently inside the pool grouped by their hash.
-    by_hash: HashMap<TxHash, Arc<PendingTransaction<T>>>,
-    /// _All_ Transactions that are currently inside the pool grouped by their identifier.
-    by_id: BTreeMap<TransactionId, Arc<PendingTransaction<T>>>,
-    /// Independent transactions that can be included directly and don't require other
-    /// transactions.
-    ///
-    /// Sorted by their scoring value.
-    independent_transactions: BTreeSet<PendingTransactionRef<T>>,
-}
-
-// === impl PendingPool ===
-
-impl<T: TransactionOrdering> PendingPool<T> {
-    /// Returns an iterator over all transactions that are _currently_ ready.
-    ///
-    /// 1. The iterator _always_ returns transaction in order: It never returns a transaction with
-    /// an unsatisfied dependency and only returns them if dependency transaction were yielded
-    /// previously. In other words: The nonces of transactions with the same sender will _always_
-    /// increase by exactly 1.
-    ///
-    /// The order of transactions which satisfy (1.) is determent by their computed priority: A
-    /// transaction with a higher priority is returned before a transaction with a lower priority.
-    ///
-    /// If two transactions have the same priority score, then the transactions which spent more
-    /// time in pool (were added earlier) are returned first.
-    ///
-    /// NOTE: while this iterator returns transaction that pool considers valid at this point, they
-    /// could potentially be become invalid at point of execution. Therefore, this iterator
-    /// provides a way to mark transactions that the consumer of this iterator considers invalid. In
-    /// which case the transaction's subgraph is also automatically marked invalid, See (1.).
-    /// Invalid transactions are skipped.
-    pub(crate) fn best(&self) -> BestTransactions<T> {
-        BestTransactions {
-            all: self.by_id.clone(),
-            independent: self.independent_transactions.clone(),
-            invalid: Default::default(),
-        }
-    }
-
-    /// Returns the ancestor the given transaction, the transaction with `nonce - 1`.
-    ///
-    /// Note: for a transaction with nonce higher than the current on chain nonce this will always
-    /// return an ancestor since all transaction in this pool are gapless.
-    fn ancestor(&self, id: &TransactionId) -> Option<(&Arc<PendingTransaction<T>>)> {
-        self.by_id.get(&id.unchecked_ancestor()?)
-    }
-
-    /// Adds a new transactions to the pending queue.
-    ///
-    /// # Panics
-    ///
-    /// if the transaction is already included
-    pub(crate) fn add_transaction(&mut self, tx: Arc<ValidPoolTransaction<T::Transaction>>) {
-        let hash = *tx.hash();
-        assert!(!self.by_hash.contains_key(&hash), "transaction already included");
-
-        let tx_id = *tx.id();
-        let submission_id = self.next_id();
-
-        let priority = self.ordering.priority(&tx.transaction);
-
-        let transaction =
-            PendingTransactionRef { submission_id, transaction: tx.clone(), priority };
-
-        // If there's _no_ ancestor in the pool, then this transaction is independent, this is
-        // guaranteed because this pool is gapless.
-        if self.ancestor(&tx_id).is_none() {
-            self.independent_transactions.insert(transaction.clone());
-        }
-
-        let transaction = Arc::new(PendingTransaction { transaction: transaction.clone() });
-
-        self.by_hash.insert(hash, transaction.clone());
-        self.by_id.insert(tx_id, transaction.clone());
-    }
-
-    /// Removes the transaction from the pool
-    pub(crate) fn remove_mined(&mut self, id: &TransactionId) {
-        if let Some(tx) = self.by_id.remove(id) {
-            self.by_hash.remove(tx.transaction.hash());
-            self.independent_transactions.remove(&tx.transaction);
-
-            // mark the next as independent if it exists
-            if let Some(unlocked) = self.by_id.get(&id.descendant()) {
-                self.independent_transactions.insert(unlocked.transaction.clone());
-            }
-        }
-    }
-
-    fn next_id(&mut self) -> u64 {
-        let id = self.id;
-        self.id = self.id.wrapping_add(1);
-        id
-    }
-
-    /// Returns the transaction for the hash if it's in the ready pool but not yet mined.
-    pub(crate) fn get(&self, hash: &TxHash) -> Option<Arc<PendingTransaction<T>>> {
-        self.by_hash.get(hash).cloned()
-    }
-}
-
-/// A transaction that is ready to be included in a block.
-pub(crate) struct PendingTransaction<T: TransactionOrdering> {
-    /// Reference to the actual transaction.
-    transaction: PendingTransactionRef<T>,
-}
-
-// == impl PendingTransaction ===
-
-impl<T: TransactionOrdering> PendingTransaction<T> {
-    /// Returns all ids this transaction satisfies.
-    pub(crate) fn id(&self) -> &TransactionId {
-        &self.transaction.transaction.transaction_id
-    }
-}
-
-impl<T: TransactionOrdering> Clone for PendingTransaction<T> {
-    fn clone(&self) -> Self {
-        Self { transaction: self.transaction.clone() }
-    }
-}
-
-/// A transaction that is ready to be included in a block.
-pub(crate) struct PendingTransactionRef<T: TransactionOrdering> {
-    /// Identifier that tags when transaction was submitted in the pool.
-    pub(crate) submission_id: u64,
-    /// Actual transaction.
-    pub(crate) transaction: Arc<ValidPoolTransaction<T::Transaction>>,
-    /// The priority value assigned by the used `Ordering` function.
-    pub(crate) priority: T::Priority,
-}
-
-impl<T: TransactionOrdering> PendingTransactionRef<T> {
-    /// The next transaction of the sender: `nonce + 1`
-    fn unlocks(&self) -> TransactionId {
-        self.transaction.transaction_id.descendant()
-    }
-
-    /// The hash for this transaction
-    fn hash(&self) -> &TxHash {
-        &self.transaction.hash()
-    }
-}
-
-impl<T: TransactionOrdering> Clone for PendingTransactionRef<T> {
-    fn clone(&self) -> Self {
-        Self {
-            submission_id: self.submission_id,
-            transaction: Arc::clone(&self.transaction),
-            priority: self.priority.clone(),
-        }
-    }
-}
-
-impl<T: TransactionOrdering> Eq for PendingTransactionRef<T> {}
-
-impl<T: TransactionOrdering> PartialEq<Self> for PendingTransactionRef<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other) == Ordering::Equal
-    }
-}
-
-impl<T: TransactionOrdering> PartialOrd<Self> for PendingTransactionRef<T> {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<T: TransactionOrdering> Ord for PendingTransactionRef<T> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        // This compares by `priority` and only if two tx have the exact same priority this compares
-        // the unique `submission_id`. This ensures that transactions with same priority are not
-        // equal, so they're not replaced in the set
-        self.priority
-            .cmp(&other.priority)
-            .then_with(|| other.submission_id.cmp(&self.submission_id))
-    }
-}
-
-/// An iterator that returns transactions that can be executed on the current state.
-pub struct BestTransactions<T: TransactionOrdering> {
-    all: BTreeMap<TransactionId, Arc<PendingTransaction<T>>>,
-    independent: BTreeSet<PendingTransactionRef<T>>,
-    invalid: HashSet<TxHash>,
-}
-
-// == impl TransactionsIterator ==
-
-impl<T: TransactionOrdering> BestTransactions<T> {
-    /// Mark the transaction and it's descendants as invalid.
-    pub(crate) fn mark_invalid(&mut self, tx: &Arc<ValidPoolTransaction<T::Transaction>>) {
-        self.invalid.insert(*tx.hash());
-    }
-}
-
-impl<T: TransactionOrdering> crate::traits::BestTransactions for BestTransactions<T> {
-    fn mark_invalid(&mut self, tx: &Self::Item) {
-        BestTransactions::mark_invalid(self, tx)
-    }
-}
-
-impl<T: TransactionOrdering> Iterator for BestTransactions<T> {
-    type Item = Arc<ValidPoolTransaction<T::Transaction>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            let best = self.independent.iter().next_back()?.clone();
-            let best = self.independent.take(&best)?;
-            let hash = best.transaction.hash();
-
-            // skip transactions that were marked as invalid
-            if self.invalid.contains(hash) {
-                debug!(
-                    target: "txpool",
-                    "[{:?}] skipping invalid transaction",
-                    hash
-                );
-                continue
-            }
-
-            // Insert transactions that just got unlocked.
-            if let Some(unlocked) = self.all.get(&best.unlocks()) {
-                self.independent.insert(unlocked.transaction.clone());
-            }
-
-            return Some(best.transaction)
-        }
-    }
-}
