+++ crates/transaction-pool/src/pool/state.rs
@@ -1,8 +1,5 @@
-use std::{
-    fmt,
-    ops::BitAndAssign,
-    sync::atomic::{AtomicU8, Ordering},
-};
+
+
 bitflags::bitflags! {
     /// Marker to assign a transaction into a sub-pool:
     #[derive(Default)]
@@ -19,31 +16,13 @@ bitflags::bitflags! {
     }
 }
 
-/// Identifier for the used Subpool
+/// Identifier for the used Sub-pool
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 #[repr(u8)]
 pub(crate) enum SubPool {
     Queued = 0,
     Pending,
     BaseFee,
-}
-
-// === impl SubPool ===
-
-impl SubPool {
-    /// Converts a u8 into the corresponding variant.
-    ///
-    /// # Panics
-    ///
-    /// If `val` does not match any variant
-    fn from_u8(val: u8) -> Self {
-        match val {
-            0 => SubPool::Queued,
-            1 => SubPool::Pending,
-            2 => SubPool::BaseFee,
-            _ => unreachable!("is shielded; qed"),
-        }
-    }
 }
 
 impl From<TxState> for SubPool {
